Move `rustc_middle::MaxUniverse` to `rustc_infer`.
Because `rust_infer` is the only crate that uses it.
